fix(macos): cross-browser menu identity, banner pane ownership, docs map
