+++ docs/docs_map.md
@@ -5209,6 +5209,7 @@ Do not edit it by hand; run `pnpm docs:map:gen`.
   - H2: First run
   - H2: Updates
   - H2: Open dashboard links
+  - H2: Import browser logins
   - H2: Choose a Gateway mode
   - H2: What the app owns
   - H2: macOS detail pages
